requirements.txt
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
Package           Version
----------------- -----------
aiosqlite         0.20.0
annotated-types   0.6.0
anyio             4.3.0
bcrypt            4.1.2
cffi              1.16.0
click             8.1.7
cryptography      42.0.5
dnspython         2.6.1
ecdsa             0.18.0
email_validator   2.1.1
exceptiongroup    1.2.0
fastapi           0.110.0
h11               0.14.0
idna              3.6
numpy             1.26.4
pandas            2.2.1
passlib           1.7.4
pip               24.0
pyasn1            0.6.0
pycparser         2.21
pydantic          2.6.4
pydantic_core     2.16.3
python-dateutil   2.9.0.post0
python-jose       3.3.0
python-multipart  0.0.9
pytz              2024.1
rsa               4.9
setuptools        58.0.4
six               1.16.0
sniffio           1.3.1
SQLAlchemy        2.0.29
starlette         0.36.3
typing_extensions 4.10.0
tzdata            2024.1
uvicorn           0.29.0
